Get Monodoc running again by zorching legacy crap and adding a null renderer
