+++ configure.in
@@ -62,43 +62,6 @@ fi
 AC_SUBST(GNOME_SHARP_LIBS)
 AM_CONDITIONAL(ENABLE_GNOME, test "x$enable_gnome" = "xyes")
 
-AC_ARG_ENABLE(gtkhtml,
-	      AC_HELP_STRING([--enable-gtkhtml], [Use GTKHTML Browser]),,
-	      enable_gtkhtml=yes)
-
-if test "x$enable_gtkhtml" = "xyes" ; then
-	PKG_CHECK_MODULES([GTKHTML_SHARP],
-			[gtkhtml-sharp-3.14],
-			[CSHARP_FLAGS="-d:GTKHTML_SHARP_3_14"],
-				[
-					PKG_CHECK_MODULES([GTKHTML_SHARP],
-						[gtkhtml-sharp-2.0],
-						[CSHARP_FLAGS=""],
-						[])
-				]
-			)
-       AC_SUBST(GTKHTML_SHARP_LIBS)
-       AC_SUBST(CSHARP_FLAGS)
-fi
-
-AM_CONDITIONAL(ENABLE_GTKHTML, test "x$enable_gtkhtml" = "xyes")
-
-AC_ARG_ENABLE(gecko,
-             AC_HELP_STRING([--enable-gecko], [Use Gecko / Xulrunner-1.8 browser libs]),,
-             enable_gecko=yes)
-
-if test "x$enable_gecko" = "xyes" ; then
-       PKG_CHECK_MODULES(GECKO_SHARP, gecko-sharp-2.0)
-
-       AC_SUBST(GECKO_SHARP_LIBS)
-
-       # get the mozilla home directory
-       MOZILLA_HOME="`$PKG_CONFIG --variable=libdir mozilla-gtkmozembed`"
-       AC_SUBST(MOZILLA_HOME)
-fi
-
-AM_CONDITIONAL(ENABLE_GECKO, test "x$enable_gecko" = "xyes")
-
 AC_ARG_ENABLE(webkit,
              AC_HELP_STRING([--enable-webkit], [Use Webkit browser]),,
              enable_webkit=yes)
@@ -110,37 +73,9 @@ fi
 
 AM_CONDITIONAL(ENABLE_WEBKIT, test "x$enable_webkit" = "xyes")
 
-# PKG_CHECK_MODULES(MONOWEBBROWSER, mono-webbrowser, enable_monowebbrowser=yes, enable_monowebbrowser=no)
-# AC_SUBST(MONOWEBBROWSER_LIBS)
-
-AC_ARG_ENABLE(monowebbrowser,
-             AC_HELP_STRING([--enable-monowebbrowser], [Use Mono.Webbrowser]),,
-             enable_monowebbrowser=yes)
-
-if test "x$enable_monowebbrowser" = "xyes" ; then
-	AC_MSG_CHECKING([for Mono.WebBrowser])
-	if test "x$GACUTIL" = "xno"
-	then
-		AC_MSG_RESULT([no])
-		if test "x$enable_monowebbrowser" = "x" ; then enable_monowebbrowser=no ; fi
-	else
-		if $GACUTIL -l Mono.WebBrowser | grep -q "Mono.WebBrowser"
-		then
-			AC_MSG_RESULT([yes])
-			enable_monowebbrowser=yes
-			MONOWEBBROWSER_LIBS=-r:Mono.WebBrowser
-		else
-			AC_MSG_RESULT([no])
-			enable_monowebbrowser=no
-		fi
-	fi
-fi
-
-AM_CONDITIONAL(ENABLE_MONOWEBBROWSER, test "x$enable_monowebbrowser" = "xyes")
-
-if test "x$enable_gtkhtml$enable_gecko$enable_monowebbrowser$enable_webkit" = "xnononono" ; then
-   AC_MSG_ERROR([no HTML display found. You need either gtkhtml, gecko, WebKit or Mono.WebBrowser])
-fi
+#if test "x$enable_webkit" = "xno" ; then
+#   AC_MSG_ERROR([no HTML display found. You need WebKit])
+#fi
 
 AC_ARG_ENABLE(debug,
 	AC_HELP_STRING([--enable-debug],
