+++ ilcontrast/Makefile.am
@@ -7,22 +7,22 @@ assemblydir = $(libdir)/ilcontrast
 desktopdir = $(datadir)/applications
 pixmapdir = $(datadir)/pixmaps
 
-if ENABLE_WEBKIT
-assembly_DATA = $(TARGET)
-bin_SCRIPTS = ilcontrast
+#if ENABLE_WEBKIT
+#assembly_DATA = $(TARGET)
+#bin_SCRIPTS = ilcontrast
+#desktop_DATA = ilcontrast.desktop
+#pixmap_DATA = ilcontrast.png
+#else
+#if ENABLE_GECKO
+#assembly_DATA = $(TARGET)
+#bin_SCRIPTS = ilcontrast
 desktop_DATA = ilcontrast.desktop
-pixmap_DATA = ilcontrast.png
-else
-if ENABLE_GECKO
-assembly_DATA = $(TARGET)
-bin_SCRIPTS = ilcontrast
-desktop_DATA = ilcontrast.desktop
-pixmap_DATA = ilcontrast.png
-else
+#pixmap_DATA = ilcontrast.png
+#else
 all:
 	echo "Install webkit-sharp or gecko-sharp to build ilContrast"
-endif
-endif
+#endif
+#endif
 
 if ENABLE_WEBKIT
 references = $(GTK_SHARP_LIBS) $(WEBKIT_SHARP_LIBS) -r:ICSharpCode.SharpZipLib.dll
@@ -38,7 +38,7 @@ resources = 							\
 
 sources =				\
 	Application.cs			\
-	ComparisonInfo.cs		\
+	#ComparisonInfo.cs		\
 	Window.cs			\
 	mono-api-diff.cs		\
 	mono-api-info.cs		\
